feat(geoutils): tree-shakeable getAdministrativeRegions function
