+++ src/geoUtils.ts
@@ -1,5 +1,3 @@
-import administrativeRegionsEl from "../data/administrative-regions-el.json";
-import administrativeRegionsEn from "../data/administrative-regions-en.json";
 import citiesEl from "../data/cities-el.json";
 import citiesEn from "../data/cities-en.json";
 import geographicRegionsEl from "../data/geographic-regions-el.json";
@@ -12,6 +10,7 @@ import taxOfficesEn from "../data/taxOffices-en.json";
 import { normalizeAndUppercaseGreekString } from "./normalizeAndUppercaseGreekString";
 import countriesEl from "../data/countries-el.json";
 import countriesEn from "../data/countries-en.json";
+import { getAdministrativeRegions } from "./getAdministrativeRegions";
 import type {
   City,
   Country,
@@ -27,16 +26,6 @@ import type {
 
 export const MOUNT_ATHOS_REGION_ID = 14;
 export const MOUNT_ATHOS_PREFECTURE_ID = 55;
-
-const administrativeRegions = {
-  el: administrativeRegionsEl,
-  en: administrativeRegionsEn,
-} as const;
-
-const administrativeRegionsWithoutMountAthos = {
-  el: administrativeRegions.el.filter(({ id }) => id !== MOUNT_ATHOS_REGION_ID),
-  en: administrativeRegions.en.filter(({ id }) => id !== MOUNT_ATHOS_REGION_ID),
-};
 
 const cities = {
   el: citiesEl,
@@ -66,36 +55,6 @@ type AdministrativeRegionsOptions = {
   includeMountAthos?: boolean;
   level?: "region" | "unit" | "municipality";
 };
-
-/**
- * Returns the administrative regions based on the provided options.
- *
- * @param {AdministrativeRegionsOptions} options - The options for locale, whether to include Mount Athos and the level area ("region" | "unit" | "municipality") to retrieve
- *
- * @returns {Region[] | RegionWithoutUnits[]} The administrative regions in the specified locale and the level area to retrieve.
- */
-export function getAdministrativeRegions({
-  locale = "el",
-  includeMountAthos = false,
-  level = "municipality",
-}: AdministrativeRegionsOptions = {}): Region[] | RegionWithoutUnits[] {
-  const regionsData = includeMountAthos
-    ? administrativeRegions[locale]
-    : administrativeRegionsWithoutMountAthos[locale];
-
-  if (level === "region") {
-    return regionsData.flatMap(({ units: _unit, ...region }) => region);
-  }
-
-  if (level === "unit") {
-    return regionsData.flatMap((region) => ({
-      ...region,
-      units: region.units.map(({ municipalities: _municipalities, ...unit }) => unit),
-    }));
-  }
-
-  return regionsData;
-}
 
 type AdministrativeRegionByIdOptions = { id: number } & AdministrativeRegionsOptions;
